eib: initial revision
Initial revision of the embedded image builder. Allows building and
chrooting into alpine linux images, configurable through config.lua,
with checkpointing support. The whole process can be run under a user
namespace instead of using sudo.

Not very well-documented or feature-rich at the moment, but this serves
as a good proof-of-concept.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>