eib: add empty "root" snapshot
After all snapshots specified in config.lua have been created, add an
empty "root" snapshot for the final rootfs to be built. The root
snapshot includes the chroot and down changes rather than polluting the
last snapshot in the config file.

When chroot'ing without a specified snapshot, use the root snapshot.
This is required since the mountpoint is dropped when the mount
namespace exits, so we always have to load a snapshot for chroot'ing.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>